requirements.txt
# Core ML dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
pytesseract>=0.3.10
Pillow>=9.5.0

# API and Web Framework
fastapi>=0.95.0
uvicorn>=0.22.0
streamlit>=1.22.0
python-multipart>=0.0.6

# MLOps and Monitoring
wandb>=0.15.0
mlflow>=2.4.0
dvc>=3.0.0
hydra-core>=1.3.2
evidently>=0.2.0
tensorboard>=2.12.0

# Cloud and Deployment
boto3>=1.26.0
sagemaker>=2.160.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0 